♻️ [refactor]: #452 Phase C-F: lexer_read_keyword.rs を分離
read_keyword の 42 テスト（Phase C: Primitive マッピング / D: 構造制御
マッピング / E: 未知キーワード平坦化 / F: 境界条件）を
`src/lexer/tests/lexer_read_keyword.rs` に分離。テスト本体で
`Token::Primitive(Primitive::Boolean(true))` 等を直接参照するため
`use super::super::token::{Primitive, Token};` を追加。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_keyword.rs: 393 行 (< 800)

Refs #452
